+++ spec/features/conversions/users_can_complete_conversion_tasks_spec.rb
@@ -32,6 +32,7 @@ RSpec.feature "Users can complete conversion tasks" do
     conditions_met
     confirm_headteacher_contact
     confirm_chair_of_governors_contact
+    confirm_incoming_trust
     main_contact
     proposed_capacity_of_the_academy
     receive_grant_payment_certificate