+++ CHANGELOG.md
@@ -17,6 +17,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   needed.
 - a new task 'confirm the chair of governors' task has been added to conversion
   projects, users must supply and indicate which contact fills this role.
+- a new 'confirm the incoming trust CEO contact' task has been added to
+  conversion projects, users need to choose the appropriate contact from those
+  available to the project, they may also have to add it
 
 ## [Release-82][release-82]